+++ Runtime/CesiumBingMapsRasterOverlay.cs
@@ -27,6 +27,8 @@ namespace CesiumForUnity
     [ReinteropNativeImplementation(
         "CesiumForUnityNative::CesiumBingMapsRasterOverlayImpl",
         "CesiumBingMapsRasterOverlayImpl.h")]
+    [AddComponentMenu("Cesium/Cesium Bing Maps Raster Overlay")]
+    [IconAttribute("Packages/com.cesium.unity/Editor/Resources/Cesium-24x24.png")]
     public partial class CesiumBingMapsRasterOverlay : CesiumRasterOverlay
     {
         [SerializeField]