+++ Runtime/CesiumCreditSystemUI.cs
@@ -21,6 +21,8 @@ namespace CesiumForUnity
     /// </summary>
     [ExecuteInEditMode]
     [RequireComponent(typeof(UIDocument))]
+    [AddComponentMenu("Cesium/Cesium Credit System UI")]
+    [IconAttribute("Packages/com.cesium.unity/Editor/Resources/Cesium-24x24.png")]
     internal class CesiumCreditSystemUI : MonoBehaviour
     {
         private CesiumCreditSystem _creditSystem;