+++ Runtime/CesiumFlyToController.cs
@@ -16,6 +16,8 @@ namespace CesiumForUnity
     /// </remarks>
     [RequireComponent(typeof(CesiumOriginShift))]
     [DisallowMultipleComponent]
+    [AddComponentMenu("Cesium/Cesium Fly To Controller")]
+    [IconAttribute("Packages/com.cesium.unity/Editor/Resources/Cesium-24x24.png")]
     public class CesiumFlyToController : MonoBehaviour
     {
         [SerializeField]