+++ Runtime/CesiumGeoreference.cs
@@ -60,6 +60,8 @@ namespace CesiumForUnity
     /// </remarks>
     [ExecuteInEditMode]
     [ReinteropNativeImplementation("CesiumForUnityNative::CesiumGeoreferenceImpl", "CesiumGeoreferenceImpl.h")]
+    [AddComponentMenu("Cesium/Cesium Georeference")]
+    [IconAttribute("Packages/com.cesium.unity/Editor/Resources/Cesium-24x24.png")]
     public partial class CesiumGeoreference : MonoBehaviour
     {
         #region Fields