+++ Runtime/CesiumIonRasterOverlay.cs
@@ -7,6 +7,8 @@ namespace CesiumForUnity
     /// A <see cref="CesiumRasterOverlay"/> that uses an IMAGERY asset from Cesium ion.
     /// </summary>
     [ReinteropNativeImplementation("CesiumForUnityNative::CesiumIonRasterOverlayImpl", "CesiumIonRasterOverlayImpl.h")]
+    [AddComponentMenu("Cesium/Cesium Ion Raster Overlay")]
+    [IconAttribute("Packages/com.cesium.unity/Editor/Resources/Cesium-24x24.png")]
     public partial class CesiumIonRasterOverlay : CesiumRasterOverlay
     {
         [SerializeField]