+++ Runtime/CesiumOriginShift.cs
@@ -25,6 +25,8 @@ namespace CesiumForUnity
     /// </remarks>
     [RequireComponent(typeof(CesiumGlobeAnchor))]
     [DisallowMultipleComponent]
+    [AddComponentMenu("Cesium/Cesium Origin Shift")]
+    [IconAttribute("Packages/com.cesium.unity/Editor/Resources/Cesium-24x24.png")]
     public class CesiumOriginShift : MonoBehaviour
     {
         void LateUpdate()