+++ Runtime/CesiumWebMapServiceRasterOverlay.cs
@@ -10,6 +10,8 @@ namespace CesiumForUnity
     [ReinteropNativeImplementation(
         "CesiumForUnityNative::CesiumWebMapServiceRasterOverlayImpl",
         "CesiumWebMapServiceRasterOverlayImpl.h")]
+    [AddComponentMenu("Cesium/Cesium Web Map Service Raster Overlay")]
+    [IconAttribute("Packages/com.cesium.unity/Editor/Resources/Cesium-24x24.png")]
     public partial class CesiumWebMapServiceRasterOverlay : CesiumRasterOverlay
     {
         [SerializeField]